create useMembersQuery hook
